new testclass FallingShould
Put all test for faling into holes or OOB into a seperate testclass FallingShould instead of having it in RobotShould.
