+++ src/test/java/inf112/skeleton/app/robottests/RobotShould.java
@@ -189,55 +189,4 @@ public class RobotShould {
 	public void rotate1CounterClockWiseOnRotate1Direction3() {
 		checkRotation(-1, CardinalDirection.WEST);
 	}
-
-	/**
-	 * MOVEMENT OUT OF BOUNDS
-	 */
-	private void checkRespawnOnOutOfBounds(int x, int y, CardinalDirection direction) {
-		robot.setDirection(direction.value);
-		robot.setX(x);
-		robot.setY(y);
-		GameMovement.move(1, robot, brd, null);
-
-		if (direction == CardinalDirection.NORTH) {
-			assertEquals(robot.getSpawnY(), robot.getY());
-		} else if (direction == CardinalDirection.EAST) {
-			assertEquals(robot.getSpawnX(), robot.getX());
-		} else if (direction == CardinalDirection.SOUTH) {
-			assertEquals(robot.getSpawnY(), robot.getY());
-		} else if (direction == CardinalDirection.WEST) {
-			assertEquals(robot.getSpawnX(), robot.getX());
-		}
-	}
-
-	@Test
-	public void respawnOnX0Y0DirectionNORTH(){
-		checkRespawnOnOutOfBounds(0, 0, CardinalDirection.NORTH);
-	}
-
-	@Test
-	public void respawnOnX0Y0DirectionWEST(){
-		checkRespawnOnOutOfBounds(0, 0, CardinalDirection.WEST);
-	}
-
-	@Test
-	public void respawnOnX11Y11DirectionSOUTH(){
-		checkRespawnOnOutOfBounds(11, 11, CardinalDirection.SOUTH);
-	}
-
-	@Test
-	public void respawnOnX11Y11DirectionEAST(){
-		checkRespawnOnOutOfBounds(11, 11, CardinalDirection.EAST);
-	}
-
-	
-	/**
-	 * FALLING IN HOLE
-	 */
-	@Test
-	public void respawnOnMovingIntoHole(){
-		Hole hole = new Hole(1);
-		brd.setTile(robotX, robotY-1, hole);
-		checkRespawnOnOutOfBounds(robotX, robotY, CardinalDirection.NORTH);
-	}
 }
